refactor: Make code into new components on Media Page
Refactored media page components, removed unused imports, and updated styles and layouts.
