+++ app/globals.css
@@ -63,7 +63,7 @@ body {
 }
 
 body {
-  min-height: 80vh;
+  min-height: 75vh;
   background-color: var(--background);
 }
